refactor: GithubBranchManager 로 infra 빈 이름 변경
