+++ backend/baton/src/test/java/touch/baton/config/infra/github/MockGithubBranchServiceConfig.java
@@ -3,7 +3,7 @@ package touch.baton.config.infra.github;
 import org.mockito.Mockito;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
-import touch.baton.infra.github.GithubBranchService;
+import touch.baton.infra.github.GithubBranchManager;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doNothing;
@@ -12,8 +12,8 @@ import static org.mockito.Mockito.doNothing;
 public abstract class MockGithubBranchServiceConfig {
 
     @Bean
-    public GithubBranchService githubBranchService() {
-        final GithubBranchService mock = Mockito.mock(GithubBranchService.class);
+    public GithubBranchManager githubBranchService() {
+        final GithubBranchManager mock = Mockito.mock(GithubBranchManager.class);
         doNothing().when(mock).createBranch(any(String.class), any(String.class));
 
         return mock;
